+++ examples/CanvasLite.js
@@ -2,14 +2,14 @@
 *   CanvasLite
 *   an html canvas implementation in pure JavaScript
 *
-*   @version 0.9.92 (2023-07-30 16:44:30)
+*   @version 0.9.92 (2023-07-30 21:49:27)
 *   https://github.com/foo123/CanvasLite
 *
 **//**
 *   CanvasLite
 *   an html canvas implementation in pure JavaScript
 *
-*   @version 0.9.92 (2023-07-30 16:44:30)
+*   @version 0.9.92 (2023-07-30 21:49:27)
 *   https://github.com/foo123/CanvasLite
 *
 **/
@@ -6397,8 +6397,23 @@ async function imagepng(type, img, width, height, metaData)
 
 function Image()
 {
-    var self = this, src = '', width = 0, height = 0, imageData = null;
+    var self = this, src = '', width = 0, height = 0, imageData = null, load;
 
+    load = function load(buffer) {
+        var imgReader = Image.Reader[Image.detectImageType(buffer)];
+        if (!imgReader) err('Image file type is not supported!');
+        imgReader(buffer)
+        .then(function(imgData) {
+            imageData = imgData;
+            width = imgData.width;
+            height = imgData.height;
+            if (self.onload) self.onload();
+        })
+        .catch(function(error) {
+            if (self.onerror) self.onerror(error);
+            else throw error;
+        });
+    };
     def(self, 'width', {
         get: function() {
             return width;
@@ -6431,32 +6446,46 @@ function Image()
         get: function() {
             return src;
         },
-        set: function(s) {
-            src = String(s);
-            if (isNode)
+        set: function(file) {
+            if ((('undefined' !== typeof ArrayBuffer) && (file instanceof ArrayBuffer))
+                || (('undefined' !== typeof Buffer) && (file instanceof Buffer)))
             {
-                require('fs').readFile(src, function(error, buffer) {
+                // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/ArrayBuffer
+                // https://nodejs.org/api/buffer.html#class-buffer
+                load(src = file);
+            }
+            else if ((('undefined' !== typeof Blob) && (file instanceof Blob))
+                || (('undefined' !== typeof Buffer) && (Buffer.Blob) && (file instanceof Buffer.Blob)))
+            {
+                // https://developer.mozilla.org/en-US/docs/Web/API/Blob
+                // https://nodejs.org/api/buffer.html#class-blob
+                (src = file).arrayBuffer()
+                .then(function(buffer) {
+                    load(buffer);
+                })
+                .catch(function(error) {
+                    if (self.onerror) self.onerror(error);
+                    else throw error;
+                })
+            }
+            else if (isNode && (('string' === typeof file) || (file instanceof String)))
+            {
+                require('fs').readFile(src = file, function(error, buffer) {
                     if (error)
                     {
                         if (self.onerror) self.onerror(error);
                         else throw error;
-                        return;
                     }
-                    var imgReader = Image.Reader[Image.detectImageType(buffer)];
-                    if (!imgReader) err('"'+src+'" image type is not supported!');
-                    imgReader(buffer)
-                    .then(function(imgData) {
-                        imageData = imgData;
-                        width = imgData.width;
-                        height = imgData.height;
-                        if (self.onload) self.onload();
-                    })
-                    .catch(function(error) {
-                        if (self.onerror) self.onerror(error);
-                        else throw error;
-                    });
+                    else
+                    {
+                        load(buffer);
+                    }
                 });
             }
+            /*else
+            {
+                err('Unsupported src property');
+            }*/
         }
     });
     self.getImageData = function() {
@@ -6482,28 +6511,28 @@ Image.Reader = {
 };
 Image.detectImageType = function(buffer) {
     // https://en.wikipedia.org/wiki/List_of_file_signatures
-    var data = new Uint8Array(buffer),
-        readByte = function(offset) {return offset < data.length ? data[offset++] : 0;};
-    if (0x89 === readByte(0)
-        && 0x50 === readByte(1)
-        && 0x4e === readByte(2)
-        && 0x47 === readByte(3)
-        && 0x0d === readByte(4)
-        && 0x0a === readByte(5)
-        && 0x1a === readByte(6)
-        && 0x0a === readByte(7)
+    var data = new Uint8Array(buffer[buffer.subarray ? 'subarray' : 'slice'](0, 8)),
+        byteAt = function(offset) {return offset < data.length ? data[offset] : 0;};
+    if (0x89 === byteAt(0)
+        && 0x50 === byteAt(1)
+        && 0x4e === byteAt(2)
+        && 0x47 === byteAt(3)
+        && 0x0d === byteAt(4)
+        && 0x0a === byteAt(5)
+        && 0x1a === byteAt(6)
+        && 0x0a === byteAt(7)
     ) return 'PNG';
-    else if (0x47 === readByte(0)
-        && 0x49 === readByte(1)
-        && 0x46 === readByte(2)
-        && 0x38 === readByte(3)
-        && (0x37 === readByte(4) || 0x39 === readByte(4))
-        && 0x61 === readByte(5)
+    else if (0x47 === byteAt(0)
+        && 0x49 === byteAt(1)
+        && 0x46 === byteAt(2)
+        && 0x38 === byteAt(3)
+        && (0x37 === byteAt(4) || 0x39 === byteAt(4))
+        && 0x61 === byteAt(5)
     ) return 'GIF';
-    else if (0xff === readByte(0)
-        && 0xd8 === readByte(1)
-        /*&& 0xff === readByte(2)
-        && 0xdb === readByte(3)*/
+    else if (0xff === byteAt(0)
+        && 0xd8 === byteAt(1)
+        /*&& 0xff === byteAt(2)
+        && 0xdb === byteAt(3)*/
     ) return 'JPG';
     return 'NOT_SUPPORTED';
 };
@@ -6565,12 +6594,18 @@ function CanvasLite(width, height)
         return await imagepng('base64', imageData.data, imageData.width, imageData.height/*, encoderOptions*/);
     };
     self.toBlob = function(cb) {
-        if (('undefined' !== typeof Blob) && cb)
+        if ('function' === typeof cb)
         {
-            // only PNG output format
-            imagepng('binary', imageData.data, imageData.width, imageData.height/*, encoderOptions*/).then(function(png) {
-                cb(new Blob(png, {type:'image/png'}));
-            }).catch(function() {});
+            var blobClass = 'undefined' !== typeof Blob ? Blob : (('undefined' !== typeof Buffer) && Buffer.Blob ? Buffer.Blob : null);
+            if (blobClass)
+            {
+                // only PNG output format
+                imagepng('binary', imageData.data, imageData.width, imageData.height)
+                .then(function(png) {
+                    cb(new blobClass([png], {type:'image/png'}));
+                })
+                .catch(NOOP);
+            }
         }
     };
     self.toPNG = async function() {